feat: 優先度選択メニューのコンポーネント化 #95
汎用性向上のため
